move external activity listener to dedicated type
